requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via httpx
boto3==1.42.76
    # via -r requirements.in
botocore==1.42.76
    # via
    #   boto3
    #   s3transfer
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
deprecation==2.1.0
    # via postgrest
gotrue==2.9.1
    # via supabase
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.25.2
    # via
    #   gotrue
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
packaging==26.0
    # via deprecation
pillow==12.1.1
    # via -r requirements.in
postgrest==0.15.1
    # via supabase
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.5
    # via
    #   gotrue
    #   postgrest
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   realtime
    #   storage3
python-dotenv==1.0.0
    # via -r requirements.in
realtime==1.0.6
    # via supabase
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via httpx
storage3==0.7.7
    # via supabase
strenum==0.4.15
    # via postgrest
supabase==2.3.5
    # via -r requirements.in
supafunc==0.3.3
    # via supabase
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   realtime
    #   storage3
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via botocore
websockets==12.0
    # via realtime
